Assembler

The assembler handles the insertion of the element matrices and element vectors into the system matrix and right hand side. While the CSC and CSR formats are the most common sparse matrix formats in practice, users might want to have optimized custom matrix formats for their specific use-case. The default assembler AssemblerSparsityPattern should be able to handle most cases in practice. To support a custom format users have to dispatch the following functions:

Missing docstring.

Missing docstring for Ferrite._assemble_inner!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ferrite.zero_out_rows!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ferrite.zero_out_columns!. Check Documenter's build log for details.

and the AbstractSparseMatrix interface for their custom matrix type.

Custom Assembler

In case the default assembler is insufficient, users can implement a custom assemblers. For this, they can create a custom type and dispatch the following functions.

Ferrite.matrix_handle — Function

matrix_handle(a::AbstractSparseAssembler)
vector_handle(a::AbstractSparseAssembler)

Return a reference to the underlying matrix/vector of the assembler.

Ferrite.vector_handle — Function

matrix_handle(a::AbstractSparseAssembler)
vector_handle(a::AbstractSparseAssembler)

Return a reference to the underlying matrix/vector of the assembler.

Missing docstring.

Missing docstring for start_assemble!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for finish_assemble!. Check Documenter's build log for details.

Ferrite.assemble! — Function

assemble!(a::Assembler, dofs, Ke)

Assembles the element matrix Ke into a.

assemble!(a::Assembler, rowdofs, coldofs, Ke)

Assembles the matrix Ke into a according to the dofs specified by rowdofs and coldofs.

assemble!(g, dofs, ge)

Assembles the element residual ge into the global residual vector g.

assemble!(A::AbstractSparseAssembler, dofs::AbstractVector{Int}, Ke::AbstractMatrix)
assemble!(A::AbstractSparseAssembler, dofs::AbstractVector{Int}, Ke::AbstractMatrix, fe::AbstractVector)

Assemble the element stiffness matrix Ke (and optional force vector fe) into the global stiffness (and force) in A, given the element degrees of freedom dofs.

This is equivalent to K[dofs, dofs] += Ke and f[dofs] += fe, where K is the global stiffness matrix and f the global force/residual vector, but more efficient.

For local elimination support the following functions might also need custom dispatches

Missing docstring.

Missing docstring for Ferrite._condense_local!. Check Documenter's build log for details.

Custom Matrix Type Sparsity Pattern

TODO create_sparsity_pattern