Dof Handler

Type definitions

Dof handlers are subtypes of AbstractDofhandler{sdim}, i.e. they are parametrized by the spatial dimension. Internally a helper struct InterpolationInfo is utilized to enforce type stability during dof distribution, because the interpolations are not available as concrete types.

Missing docstring.

Missing docstring for Ferrite.AbstractDofhandler. Check Documenter's build log for details.

Ferrite.InterpolationInfo — Type

InterpolationInfo

Gathers all the information needed to distribute dofs for a given interpolation. Note that this cache is of the same type no matter the interpolation: the purpose is to make dof-distribution type-stable.

Ferrite.PathOrientationInfo — Type

Orientation information for 1D entities.

The orientation for 1D entities is defined by the indices of the grid nodes associated to the vertices. To give an example, the oriented path 1 –-> 2 is called regular, indicated by regular=true, while the oriented path 2 –-> 1 is called inverted, indicated by regular=false.

Ferrite.SurfaceOrientationInfo — Type

Orientation information for 2D entities. Such an entity can be possibly flipped (i.e. the defining vertex order is reverse to the spanning vertex order) and the vertices can be rotated against each other. Take for example the faces 1–-2 2–-3 | A | | B | 4–-3 1–-4 which are rotated against each other by 90° (shift index is 1) or the faces 1–-2 2–-1 | A | | B | 4–-3 4–-3 which are flipped against each other. Any combination of these can happen. The combination to map this local face to the defining face is encoded with this data structure via $rotate \circ flip$ where the rotation is indiced by the shift index. !!!NOTE TODO implement me.

Internal API

The main entry point for dof distribution is __close!.

Ferrite.find_field — Method

find_field(dh::DofHandler, field_name::Symbol)::NTuple{2,Int}

Return the index of the field with name field_name in a DofHandler. The index is a NTuple{2,Int}, where the 1st entry is the index of the FieldHandler within which the field was found and the 2nd entry is the index of the field within the FieldHandler.

Note

Always finds the 1st occurence of a field within DofHandler.

See also: find_field(fh::FieldHandler, field_name::Symbol), _find_field(fh::FieldHandler, field_name::Symbol).

Missing docstring.

Missing docstring for Ferrite.find_field(fh::FieldHandler, field_name::Symbol)::Int. Check Documenter's build log for details.

Ferrite._find_field — Method

_find_field(fh::FieldHandler, field_name::Symbol)::Int

Return the index of the field with name field_name in the FieldHandler fh. Return nothing if the field is not found.

See also: find_field(dh::DofHandler, field_name::Symbol), find_field(fh::FieldHandler, field_name::Symbol).

Missing docstring.

Missing docstring for Ferrite.__close(::DofHandler). Check Documenter's build log for details.

Ferrite._close_fieldhandler! — Function

_close_fieldhandler!(dh::DofHandler{sdim}, fh::FieldHandler, fh_index::Int, nextdof::Int, vertexdicts::VD, edgedicts::ED, facedicts::FD)

Main entry point to distribute dofs for a single FieldHandler on its subdomain.

Ferrite._close_fieldhandler_on_cell! — Function

_close_fieldhandler_on_cell!(dh::DofHandler{sdim}, cell::C, fh::FieldHandler, vdim::Int, ip_info::InterpolationInfo, nextdof::Int, vertexdict::VD, edgedict::ED, facedict::FD)

Main entry point to distribute dofs for a single FieldHandler on a given cell.

Missing docstring.

Missing docstring for Ferrite._check_cellset_intersections. Check Documenter's build log for details.

Ferrite.dof_correction! — Method

For interpolations with more than one interior dof and some dofs associated with the edge will have non-matching evaluations on these edges if the orientations do not match. Think e.g. about the following nodal interpolation example with two edges with three interior dofs on each edge: +–––––-+ | A | +–1–2–3->+ local edge on element A

–––––> global edge

+<-3–2–1–+ local edge on element B | B | +–––––-+ Here the dofs will be locally assigned as indicated by 1,2,3 and we see a mismatch between the dofs 1 and 3. For most scalar-valued interpolations we can simply compensate this by reverting the dof assignment on the edge interior.

In addition, we also have to preverse the ordering at each dof location.

For more details we refer to [1] as we follow the methodology described therein.

[1] Scroggs, M. W., Dokken, J. S., Richardson, C. N., & Wells, G. N. (2022). Construction of arbitrary order finite element degree-of-freedom maps on polygonal and polyhedral cell meshes. ACM Transactions on Mathematical Software (TOMS), 48(2), 1-23.

!!!TODO Citation via DocumenterCitations.jl.

!!!TODO Investigate if we can somehow pass the interpolation into this function in a typestable way.

Ferrite.dof_correction! — Method

!!!NOTE TODO implement me.

For more details we refer to [1] as we follow the methodology described therein.

[1] Scroggs, M. W., Dokken, J. S., Richardson, C. N., & Wells, G. N. (2022). Construction of arbitrary order finite element degree-of-freedom maps on polygonal and polyhedral cell meshes. ACM Transactions on Mathematical Software (TOMS), 48(2), 1-23.

!!!TODO citation via software.

!!!TODO Investigate if we can somehow pass the interpolation into this function in a typestable way.

Ferrite._dof_correction! — Method

_dof_correction!(cell_dofs::Vector{Int}, dofs::StepRange{Int,Int})

Insert dofs in order.

Missing docstring.

Missing docstring for Ferrite.add_vertex_dofs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ferrite.add_face_dofs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ferrite.add_edge_dofs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Ferrite.add_cell_dofs. Check Documenter's build log for details.