Interpolation

Ferrite.Interpolation — Type

Interpolation{ref_dim, ref_shape, order}()

Return an Interpolation on a ref_dim-dimensional reference shape (see AbstractRefShape) ref_shape and order order. order corresponds to the order of the interpolation. The interpolation is used to define shape functions to interpolate a function between nodes.

The following interpolations are implemented:

Lagrange{1,RefCube,1}

Lagrange{1,RefCube,2}

Lagrange{2,RefCube,1}

Lagrange{2,RefCube,2}

Lagrange{2,RefTetrahedron,1}

Lagrange{2,RefTetrahedron,2}

Lagrange{2,RefTetrahedron,3}

Lagrange{2,RefTetrahedron,4}

Lagrange{2,RefTetrahedron,5}

BubbleEnrichedLagrange{2,RefTetrahedron,1}

CrouzeixRaviart{2,1}

Lagrange{3,RefCube,1}

Lagrange{3,RefCube,2}

Lagrange{3,RefTetrahedron,1}

Lagrange{3,RefTetrahedron,2}

Lagrange{3,RefPrism,1}

Lagrange{3,RefPrism,2}

Serendipity{2,RefCube,2}

Serendipity{3,RefCube,2}

Examples

julia> ip = Lagrange{2,RefTetrahedron,2}()
Ferrite.Lagrange{2,Ferrite.RefTetrahedron,2}()

julia> getnbasefunctions(ip)
6

Ferrite.getnbasefunctions — Function

Ferrite.getnbasefunctions(ip::Interpolation)

Return the number of base functions for the interpolation ip.

Ferrite.getdim — Function

Ferrite.getdim(::Interpolation)

Return the dimension of the reference element for a given interpolation.

Ferrite.getrefshape — Function

Ferrite.getrefshape(::Interpolation)::AbstractRefShape

Return the reference element shape of the interpolation.

Ferrite.getorder — Function

Ferrite.getorder(::Interpolation)

Return order of the interpolation.