Assembly

Type definitions

Ferrite.COOAssembler — Type

struct COOAssembler{Tv, Ti}

This assembler creates a COO (coordinate format) representation of a sparse matrix during assembly and converts it into a SparseMatrixCSC{Tv, Ti} on finalization.

Ferrite.CSCAssembler — Type

Assembler for sparse matrix with CSC storage type.

Ferrite.SymmetricCSCAssembler — Type

Assembler for symmetric sparse matrix with CSC storage type.

Utility functions

Ferrite.matrix_handle — Function

matrix_handle(a::AbstractAssembler)
vector_handle(a::AbstractAssembler)

Return a reference to the underlying matrix/vector of the assembler used during assembly operations.

Ferrite.vector_handle — Function

matrix_handle(a::AbstractAssembler)
vector_handle(a::AbstractAssembler)

Return a reference to the underlying matrix/vector of the assembler used during assembly operations.

Ferrite._sortdofs_for_assembly! — Function

_sortdofs_for_assembly!(permutation::Vector{Int}, sorteddofs::Vector{Int}, dofs::AbstractVector)

Sorts the dofs into a separate buffer and returns it together with a permutation vector.

Ferrite.sortperm2! — Function

sortperm2!(data::AbstractVector, permutation::AbstractVector)

Sort the input vector inplace and compute the corresponding permutation.