Elements and cells

Type definitions

Elements or cells are subtypes of AbstractCell{dim,N,M}. They are parametrized by the dimension of their nodes via dim, the number of nodes N and the number of faces M.

Required methods to implement for all subtypes of AbstractCell to define a new element

Ferrite.vertices — Method

Ferrite.vertices(::AbstractCell)

Returns a tuple with the node indices (of the nodes in a grid) for each vertex in a given cell. This function induces the VertexIndex, where the second index corresponds to the local index into this tuple.

Ferrite.edges — Method

Ferrite.edges(::AbstractCell)

Returns a tuple of 2-tuples containing the ordered node indices (of the nodes in a grid) corresponding to the vertices that define an oriented edge. This function induces the EdgeIndex, where the second index corresponds to the local index into this tuple.

Note that the vertices are sufficient to define an edge uniquely.

Ferrite.faces — Method

Ferrite.faces(::AbstractCell)

Returns a tuple of n-tuples containing the ordered node indices (of the nodes in a grid) corresponding to the vertices that define an oriented face. This function induces the FaceIndex, where the second index corresponds to the local index into this tuple.

Note that the vertices are sufficient to define a face uniquely.

Ferrite.default_interpolation — Method

Ferrite.default_interpolation(::AbstractCell)::Interpolation

Returns the interpolation which defines the geometry of a given cell.

Common utilities and definitions when working with grids internally.

Ferrite.BoundaryIndex — Type

Abstract type which is used as identifier for faces, edges and verices

Ferrite.boundaryfunction — Method

boundaryfunction(::Type{<:BoundaryIndex})

Helper function to dispatch on the correct entity from a given boundary index.

Ferrite.get_coordinate_eltype — Method

Return the number type of the nodal coordinates.

Ferrite.get_coordinate_eltype — Method

get_coordinate_eltype(::Node)

Get the data type of the components of the nodes coordinate.

Ferrite.toglobal — Function

toglobal(grid::AbstractGrid, vertexidx::VertexIndex) -> Int
toglobal(grid::AbstractGrid, vertexidx::Vector{VertexIndex}) -> Vector{Int}

This function takes the local vertex representation (a VertexIndex) and looks up the unique global id (an Int).

Ferrite.sortface — Function

sortface(face::Tuple{Int})
sortface(face::Tuple{Int,Int})
sortface(face::Tuple{Int,Int,Int})
sortface(face::Tuple{Int,Int,Int,Int})

Returns the unique representation of a face. Here the unique representation is the sorted node index tuple. Note that in 3D we only need indices to uniquely identify a face, so the unique representation is always a tuple length 3.

Ferrite.sortedge — Function

sortedge(edge::Tuple{Int,Int})

Returns the unique representation of an edge and its orientation. Here the unique representation is the sorted node index tuple. The orientation is true if the edge is not flipped, where it is false if the edge is flipped.

Ferrite.getfaceedges — Function

getfaceedges(grid::AbstractGrid, face::FaceIndex)
getfaceedges(cell::AbstractCell, face::FaceIndex)

Returns the edges represented as Set{EdgeIndex} in a given face represented as FaceIndex.

julia> using Ferrite; using Ferrite: getfaceedges

julia> grid = generate_grid(Tetrahedron, (2,1,1));

julia> getfaceedges(grid, FaceIndex(4,2))
Set{EdgeIndex} with 3 elements:
  EdgeIndex((4, 4))
  EdgeIndex((4, 5))
  EdgeIndex((4, 1))

Ferrite.getfacevertices — Function

getfacevertices(grid::AbstractGrid, face::FaceIndex)
getfacevertices(cell::AbstractCell, face::FaceIndex)

Returns the vertices represented as Set{VertexIndex} in a given face represented as FaceIndex.

julia> using Ferrite; using Ferrite: getfacevertices

julia> grid = generate_grid(Tetrahedron, (2,1,1));

julia> getfacevertices(grid, FaceIndex(4,2))
Set{VertexIndex} with 3 elements:
  VertexIndex((4, 2))
  VertexIndex((4, 4))
  VertexIndex((4, 1))

Ferrite.getedgevertices — Function

getedgevertices(grid::AbstractGrid, edge::EdgeIndex)
getedgevertices(cell::AbstractCell, edge::EdgeIndex)

Returns the vertices represented as Set{VertexIndex} in a given edge represented as EdgeIndex.

julia> using Ferrite; using Ferrite: getedgevertices

julia> grid = generate_grid(Tetrahedron, (2,1,1));

julia> getedgevertices(grid, EdgeIndex(4,2))
Set{EdgeIndex} with 2 elements:
  VertexIndex((4, 2))
  VertexIndex((4, 3))

Ferrite.getfaceinstances — Function

getfaceinstances(grid::AbstractGrid, topology::ExclusiveTopology, face::FaceIndex)

Returns all the faces as Set{FaceIndex} that share all their vertices with a given face represented as FaceIndex. The returned set includes the input face.

julia> using Ferrite; using Ferrite: getfaceinstances

julia> grid = generate_grid(Tetrahedron, (2,1,1));

julia> topology = ExclusiveTopology(grid);

julia> getfaceinstances(grid, topology, FaceIndex(4,2))
Set{FaceIndex} with 2 elements:
FaceIndex((6, 4))
FaceIndex((4, 2))

Ferrite.getedgeinstances — Function

getedgeinstances(grid::AbstractGrid, topology::ExclusiveTopology, edge::EdgeIndex)

Returns all the edges as Set{EdgeIndex} that share all their vertices with a given edge represented as EdgeIndex. The returned set includes the input edge.

julia> using Ferrite; using Ferrite: getedgeinstances

julia> grid = generate_grid(Tetrahedron, (2,1,1));

julia> topology = ExclusiveTopology(grid);

julia> getedgeinstances(grid, topology, EdgeIndex(4,2))
Set{EdgeIndex} with 3 elements:
EdgeIndex((4, 2))
EdgeIndex((9, 6))
EdgeIndex((7, 6))

Ferrite.getvertexinstances — Function

getvertexinstances(grid::AbstractGrid, topology::ExclusiveTopology, vertex::EdgeIndex)

Returns all the vertices as Set{::VertexIndex} that use a given vertex represented as VertexIndex in all cells. The returned set includes the input vertex.

julia> using Ferrite; using Ferrite: getvertexinstances

julia> grid = generate_grid(Tetrahedron,(2,1,1));

julia> topology = ExclusiveTopology(grid);

julia> getvertexinstances(grid, topology, VertexIndex(4,2))
Set{VertexIndex} with 8 elements:
VertexIndex((7, 4))
VertexIndex((10, 4))
VertexIndex((12, 4))
VertexIndex((6, 3))
VertexIndex((4, 2))
VertexIndex((9, 4))
VertexIndex((11, 4))
VertexIndex((8, 4))

Ferrite.filterfaces — Function

filterfaces(grid::AbstractGrid, faces::Set{FaceIndex}, f::Function; all::Bool=true)

Returns the faces in faces that satisfy f as a Set{FaceIndex}. all=true implies that f(x) must return true for all nodal coordinates x on the face if the face should be added to the set, otherwise it suffices that f(x) returns true for one node.

julia> using Ferrite; using Ferrite: filterfaces

julia> grid = generate_grid(Tetrahedron, (2,2,2));

julia> topology = ExclusiveTopology(grid);

julia> addboundaryfaceset!(grid, topology, "b", x -> true);

julia> filterfaces(grid, grid.facesets["b"], x -> x[3] ≈ -1)
Set{FaceIndex} with 8 elements:
FaceIndex((7, 1))
FaceIndex((3, 1))
FaceIndex((21, 1))
FaceIndex((13, 1))
FaceIndex((19, 1))
FaceIndex((15, 1))
FaceIndex((1, 1))
FaceIndex((9, 1))

Ferrite.filteredges — Function

filteredges(grid::AbstractGrid, edges::Set{EdgeIndex}, f::Function; all::Bool=true)

Returns the edges in edges that satisfy f as a Set{EdgeIndex}. all=true implies that f(x) must return true for all nodal coordinates x on the face if the face should be added to the set, otherwise it suffices that f(x) returns true for one node.

julia> using Ferrite; using Ferrite: filteredges

julia> grid = generate_grid(Tetrahedron, (1,1,1));

julia> topology = ExclusiveTopology(grid);

julia> addboundaryedgeset!(grid, topology, "b", x -> true);

julia> filteredges(grid, grid.edgesets["b"], x -> x[3] ≈ -1)
Set{EdgeIndex} with 8 elements:
EdgeIndex((1, 2))
EdgeIndex((3, 2))
EdgeIndex((4, 3))
EdgeIndex((1, 3))
EdgeIndex((3, 3))
EdgeIndex((1, 1))
EdgeIndex((3, 1))
EdgeIndex((2, 3))

Ferrite.filtervertices — Function

filtervertices(grid::AbstractGrid, vertices::Set{VertexIndex}, f::Function; all::Bool=true)

Returns the vertices in vertices that satisfy f as a Set{VertexIndex}. all=true implies that f(x) must return true for all nodal coordinates x on the face if the face should be added to the set, otherwise it suffices that f(x) returns true for one node.

julia> using Ferrite; using Ferrite: filtervertices

julia> grid = generate_grid(Tetrahedron,(1,1,1));

julia> topology = ExclusiveTopology(grid);

julia> addboundaryvertexset!(grid, topology, "b", x -> true);

julia> filtervertices(grid, grid.vertexsets["b"], x -> x[3] ≈ -1)
Set{VertexIndex} with 12 elements:
VertexIndex((2, 3))
VertexIndex((4, 3))
VertexIndex((4, 1))
VertexIndex((3, 3))
VertexIndex((3, 2))
VertexIndex((1, 1))
VertexIndex((2, 1))
VertexIndex((3, 1))
VertexIndex((1, 3))
VertexIndex((5, 1))
VertexIndex((1, 2))
VertexIndex((6, 1))