Assembly

Ferrite.start_assemble — Function

start_assemble([N=0]) -> Assembler

Call before starting an assembly.

Returns an Assembler type that is used to hold the intermediate data before an assembly is finished.

Ferrite.assemble! — Function

assemble!(a, Ke, edof)

Assembles the element matrix Ke into a.

assemble!(g, ge, edof)

Assembles the element residual ge into the global residual vector g.

Ferrite.end_assemble — Function

end_assemble(a::Assembler) -> K

Finalizes an assembly. Returns a sparse matrix with the assembled values.

Ferrite.create_sparsity_pattern — Function

create_sparsity_pattern(dh::DofHandler)

Create the sparsity pattern corresponding to the degree of freedom numbering in the DofHandler. Return a SparseMatrixCSC with stored values in the correct places.

See the Sparsity Pattern section of the manual.

Ferrite.create_symmetric_sparsity_pattern — Function

create_symmetric_sparsity_pattern(dh::DofHandler)

Create the symmetric sparsity pattern corresponding to the degree of freedom numbering in the DofHandler by only considering the upper triangle of the matrix. Return a Symmetric{SparseMatrixCSC}.

See the Sparsity Pattern section of the manual.