Degrees of Freedom

Degrees of freedom (dofs) are distributed by the DofHandler or the MixedDofHandler.

Ferrite.DofHandler — Type

DofHandler(grid::Grid)

Construct a DofHandler based on grid.

Operates slightly faster than MixedDofHandler. Supports:

Grids with a single concrete cell type.

One or several fields on the whole domaine.

Ferrite.MixedDofHandler — Type

MixedDofHandler(grid::Grid)

Construct a MixedDofHandler based on grid. Supports:

Grids with or without concrete element type (E.g. "mixed" grids with several different element types.)

One or several fields, which can live on the whole domain or on subsets of the Grid.

Adding fields to the DofHandlers

Base.push! — Method

push!(dh::AbstractDofHandler, name::Symbol, dim::Int[, ip::Interpolation])

Add a dim-dimensional Field called name which is approximated by ip to dh.

The field is added to all cells of the underlying grid. In case no interpolation ip is given, the default interpolation of the grid's celltype is used. If the grid uses several celltypes, push!(dh::MixedDofHandler, fh::FieldHandler) must be used instead.

Base.push! — Method

push!(dh::MixedDofHandler, fh::FieldHandler)

Add all fields of the FieldHandler fh to dh.

Ferrite.Field — Type

Field(name::Symbol, interpolation::Interpolation, dim::Int)

Construct dim-dimensional Field called name which is approximated by interpolation.

The interpolation is used for distributing the degrees of freedom.

Ferrite.FieldHandler — Type

FieldHandler(fields::Vector{Field}, cellset::Set{Int})

Construct a FieldHandler based on an array of Fields and assigns it a set of cells.

A FieldHandler must fullfill the following requirements:

All Cells in cellset are of the same type.

Each field only uses a single interpolation on the cellset.

Each cell belongs only to a single FieldHandler, i.e. all fields on a cell must be added within the same FieldHandler.

Notice that a FieldHandler can hold several fields.

Ferrite.close! — Method

close!(dh::AbstractDofHandler)

Closes dh and creates degrees of freedom for each cell.

If there are several fields, the dofs are added in the following order: For a MixedDofHandler, go through each FieldHandler in the order they were added. For each field in the FieldHandler or in the DofHandler (again, in the order the fields were added), create dofs for the cell. This means that dofs on a particular cell, the dofs will be numbered according to the fields; first dofs for field 1, then field 2, etc.

Common methods

Ferrite.ndofs — Function

ndofs(dh::AbstractDofHandler)

Return the number of degrees of freedom in dh

Ferrite.dof_range — Function

dof_range(dh:DofHandler, field_name)

Return the local dof range for field_name. Example:

julia> grid = generate_grid(Triangle, (3, 3))
Grid{2, Triangle, Float64} with 18 Triangle cells and 16 nodes

julia> dh = DofHandler(grid); push!(dh, :u, 3); push!(dh, :p, 1); close!(dh);

julia> dof_range(dh, :u)
1:9

julia> dof_range(dh, :p)
10:12

Ferrite.nfields — Method

nfields(dh::MixedDofHandler)

Returns the number of unique fields defined.

Ferrite.getfieldnames — Method

getfieldnames(dh::MixedDofHandler)

Returns the union of all the fields. Can be used as an iterable over all the fields in the problem.

Ferrite.getfielddim — Method

getfielddim(dh::MixedDofHandler, name::Symbol)

Returns the dimension of a specific field, given by name. Note that it will return the dimension of the first field found among the FieldHandlers.

CellIterator

Ferrite.CellIterator — Type

CellIterator(grid::Grid)
CellIterator(dh::DofHandler)
CellIterator(mdh::MixedDofHandler)

Return a CellIterator to conveniently loop over all the cells in a grid.

Examples

for cell in CellIterator(dh)      # dh::DofHandler
    coords = getcoordinates(cell) # get the coordinates
    dofs = celldofs(cell)         # get the dofs for this cell
    reinit!(cv, cell)             # reinit! the FE-base with a CellIterator
end

Here, cell::CellIterator. Looking at a specific cell (instead of looping over all), e.g. nr 10, can be done by

cell = CellIterator(dh)     # Refers to cell nr. 1 upon creation
reinit!(cell, 10)           # Update to cell nr. 10
dofs = celldofs(cell)       # Get the dofs for cell nr. 10