Degrees of Freedom

Degrees of freedom (dofs) are distributed by the DofHandler.

Ferrite.DofHandler — Type

DofHandler(grid::Grid)

Construct a DofHandler based on grid. Supports:

Grids with or without concrete element type (E.g. "mixed" grids with several different element types.)

One or several fields, which can live on the whole domain or on subsets of the Grid.

Adding fields to the DofHandlers

Ferrite.add! — Method

add!(dh::DofHandler, name::Symbol, ip::Interpolation)

Add a field called name approximated by ip to the DofHandler dh.

The field is added to all cells of the underlying grid, use SubDofHandlers if the grid contains multiple cell types, or to add the field to subset of all the cells.

Ferrite.close! — Method

close!(dh::AbstractDofHandler)

Closes dh and creates degrees of freedom for each cell.

Dof renumbering

Ferrite.renumber! — Function

renumber!(dh::AbstractDofHandler, order)
renumber!(dh::AbstractDofHandler, ch::ConstraintHandler, order)

Renumber the degrees of freedom in the DofHandler and/or ConstraintHandler according to the ordering order.

order can be given by one of the following options:

A permutation vector perm::AbstractVector{Int} such that dof i is renumbered to perm[i].

DofOrder.FieldWise() for renumbering dofs field wise.

DofOrder.ComponentWise() for renumbering dofs component wise.

DofOrder.Ext{T} for "external" renumber permutations, see documentation for DofOrder.Ext for details.

Warning

The dof numbering in the DofHandler and ConstraintHandler must always be consistent. It is therefore necessary to either renumber before creating the ConstraintHandler in the first place, or to renumber the DofHandler and the ConstraintHandler together.

Ferrite.DofOrder.FieldWise — Type

DofOrder.FieldWise()
DofOrder.FieldWise(target_blocks::Vector{Int})

Dof order passed to renumber! to renumber global dofs field wise resulting in a globally blocked system.

The default behavior is to group dofs of each field into their own block, with the same order as in the DofHandler. This can be customized by passing a vector of the same length as the total number of fields in the DofHandler (see getfieldnames(dh)) that maps each field to a "target block": to renumber a DofHandler with three fields :u, :v, :w such that dofs for :u and :w end up in the first global block, and dofs for :v in the second global block use DofOrder.FieldWise([1, 2, 1]).

This renumbering is stable such that the original relative ordering of dofs within each target block is maintained.

Ferrite.DofOrder.ComponentWise — Type

DofOrder.ComponentWise()
DofOrder.ComponentWise(target_blocks::Vector{Int})

Dof order passed to renumber! to renumber global dofs component wise resulting in a globally blocked system.

The default behavior is to group dofs of each component into their own block, with the same order as in the DofHandler. This can be customized by passing a vector of length ncomponents that maps each component to a "target block" (see DofOrder.FieldWise for details).

This renumbering is stable such that the original relative ordering of dofs within each target block is maintained.

Common methods

Ferrite.ndofs — Function

ndofs(dh::AbstractDofHandler)

Return the number of degrees of freedom in dh

Ferrite.ndofs_per_cell — Function

ndofs_per_cell(dh::AbstractDofHandler[, cell::Int=1])

Return the number of degrees of freedom for the cell with index cell.

See also ndofs.

Ferrite.dof_range — Function

dof_range(sdh::SubDofHandler, field_idx::Int)
dof_range(sdh::SubDofHandler, field_name::Symbol)
dof_range(dh:DofHandler, field_name::Symbol)

Return the local dof range for a given field. The field can be specified by its name or index, where field_idx represents the index of a field within a SubDofHandler and field_idxs is a tuple of the SubDofHandler-index within the DofHandler and the field_idx.

Note

The dof_range of a field can vary between different SubDofHandlers. Therefore, it is advised to use the field_idxs or refer to a given SubDofHandler directly in case several SubDofHandlers exist. Using the field_name will always refer to the first occurence of field within the DofHandler.

Example:

julia> grid = generate_grid(Triangle, (3, 3))
Grid{2, Triangle, Float64} with 18 Triangle cells and 16 nodes

julia> dh = DofHandler(grid); add!(dh, :u, 3); add!(dh, :p, 1); close!(dh);

julia> dof_range(dh, :u)
1:9

julia> dof_range(dh, :p)
10:12

julia> dof_range(dh, (1,1)) # field :u
1:9

julia> dof_range(dh.subdofhandlers[1], 2) # field :p
10:12

Ferrite.celldofs — Function

celldofs(dh::AbstractDofHandler, i::Int)

Return a vector with the degrees of freedom that belong to cell i.

See also celldofs!.

Ferrite.celldofs! — Function

celldofs!(global_dofs::Vector{Int}, dh::AbstractDofHandler, i::Int)

Store the degrees of freedom that belong to cell i in global_dofs.

See also celldofs.

Grid iterators

Ferrite.CellCache — Type

CellCache(grid::Grid)
CellCache(dh::AbstractDofHandler)

Create a cache object with pre-allocated memory for the nodes, coordinates, and dofs of a cell. The cache is updated for a new cell by calling reinit!(cache, cellid) where cellid::Int is the cell id.

Struct fields of CellCache

cc.nodes :: Vector{Int}: global node ids

cc.coords :: Vector{<:Vec}: node coordinates

cc.dofs :: Vector{Int}: global dof ids (empty when constructing the cache from a grid)

Methods with CellCache

reinit!(cc, i): reinitialize the cache for cell i

cellid(cc): get the cell id of the currently cached cell

getnodes(cc): get the global node ids of the cell

get_cell_coordinates(cc): get the coordinates of the cell

celldofs(cc): get the global dof ids of the cell

reinit!(fev, cc): reinitialize CellValues or FaceValues

See also CellIterator.

Ferrite.CellIterator — Type

CellIterator(grid::Grid, cellset=1:getncells(grid))
CellIterator(dh::AbstractDofHandler, cellset=1:getncells(dh))

Create a CellIterator to conveniently iterate over all, or a subset, of the cells in a grid. The elements of the iterator are CellCaches which are properly reinit!ialized. See CellCache for more details.

Looping over a CellIterator, i.e.:

for cc in CellIterator(grid, cellset)
    # ...
end

is thus simply convenience for the following equivalent snippet:

cc = CellCache(grid)
for idx in cellset
    reinit!(cc, idx)
    # ...
end

Warning

CellIterator is stateful and should not be used for things other than for-looping (e.g. broadcasting over, or collecting the iterator may yield unexpected results).

Ferrite.FaceCache — Type

FaceCache(grid::Grid)
FaceCache(dh::AbstractDofHandler)

Create a cache object with pre-allocated memory for the nodes, coordinates, and dofs of a cell suitable for looping over faces in a grid. The cache is updated for a new face by calling reinit!(cache, fi::FaceIndex).

Methods with fc::FaceCache

reinit!(fc, fi): reinitialize the cache for face fi::FaceIndex

cellid(fc): get the current cellid (faceindex(fc)[1])

getnodes(fc): get the global node ids of the cell

get_cell_coordinates(fc): get the coordinates of the cell

celldofs(fc): get the global dof ids of the cell

reinit!(fv, fc): reinitialize FaceValues

See also FaceIterator.

Ferrite.FaceIterator — Type

FaceIterator(gridordh::Union{Grid,AbstractDofHandler}, faceset::Set{FaceIndex})

Create a FaceIterator to conveniently iterate over the faces in faceset. The elements of the iterator are FaceCaches which are properly reinit!ialized. See FaceCache for more details.

Looping over a FaceIterator, i.e.:

for fc in FaceIterator(grid, faceset)
    # ...
end

is thus simply convenience for the following equivalent snippet: ```julia fc = FaceCache(grid) for faceindex in faceset reinit!(fc, faceindex) # ... end