Postprocessing

Project to nodes

Ferrite.L2Projector — Type

L2Projector(func_ip::Interpolation, grid::AbstractGrid; kwargs...)

Create an L2Projector used for projecting quadrature data. func_ip is the function interpolation used for the projection and grid the grid over which the projection is applied.

Keyword arguments:

qr_lhs: quadrature for the left hand side. Defaults to a quadrature which exactly integrates a mass matrix with func_ip as the interpolation.

set: element set over which the projection applies. Defaults to all elements in the grid.

geom_ip: geometric interpolation. Defaults to the default interpolation for the grid.

The L2Projector acts as the integrated left hand side of the projection equation: Find projection $u \in L_2(\Omega)$ such that

\[\int v u \ \mathrm{d}\Omega = \int v f \ \mathrm{d}\Omega \quad \forall v \in L_2(\Omega),\]

where $f$ is the data to project.

Use project to integrate the right hand side and solve for the system.

Ferrite.project — Function

project(proj::L2Projector, vals, qr_rhs::QuadratureRule)

Makes a L2 projection of data vals to the nodes of the grid using the projector proj (see L2Projector).

project integrates the right hand side, and solves the projection $u$ from the following projection equation: Find projection $u \in L_2(\Omega)$ such that

\[\int v u \ \mathrm{d}\Omega = \int v f \ \mathrm{d}\Omega \quad \forall v \in L_2(\Omega),\]

where $f$ is the data to project, i.e. vals.

The data vals should be a vector, with length corresponding to number of elements, of vectors, with length corresponding to number of quadrature points per element, matching the number of points in qr_rhs. Alternatively, vals can be a matrix, with number of columns corresponding to number of elements, and number of rows corresponding to number of points in qr_rhs. Example (scalar) input data:

vals = [
    [0.44, 0.98, 0.32], # data for quadrature point 1, 2, 3 of element 1
    [0.29, 0.48, 0.55], # data for quadrature point 1, 2, 3 of element 2
    # ...
]

or equivalent in matrix form:

vals = [
    0.44 0.29 # ...
    0.98 0.48 # ...
    0.32 0.55 # ...
]

Supported data types to project are Numbers and AbstractTensors.

The order of the returned data correspond to the order of the L2Projectors internal DofHandler. To export the result, use vtk_point_data(vtk, proj, projected_data).

Postprocessing

Ferrite.PointEvalHandler — Type

PointEvalHandler(grid::Grid, points::AbstractVector{Vec{dim,T}}; kwargs...) where {dim, T}

The PointEvalHandler can be used for function evaluation in arbitrary points in the domain – not just in quadrature points or nodes.

The PointEvalHandler takes the following keyword arguments:

search_nneighbors: How many nodes should be found in the nearest neighbor search for each point. Usually there is no need to change this setting. Default value: 3.

warn: Show a warning if a point is not found. Default value: true.

The constructor takes a grid and a vector of coordinates for the points. The PointEvalHandler computes i) the corresponding cell, and ii) the (local) coordinate within the cell, for each point. The fields of the PointEvalHandler are:

cells::Vector{Union{Int,Nothing}}: vector with cell IDs for the points, with nothing for points that could not be found.

local_coords::Vector{Union{Vec,Nothing}}: vector with the local coordinates (i.e. coordinates in the reference configuration) for the points, with nothing for points that could not be found.

There are two ways to use the PointEvalHandler to evaluate functions:

evaluate_at_points: can be used when the function is described by i) a dh::DofHandler + uh::Vector (for example the FE-solution), or ii) a p::L2Projector + ph::Vector (for projected data).

Iteration with PointIterator + PointValues: can be used for more flexible evaluation in the points, for example to compute gradients.

Ferrite.evaluate_at_points — Function

evaluate_at_points(ph::PointEvalHandler, dh::AbstractDofHandler, dof_values::Vector{T}, [fieldname::Symbol]) where T
evaluate_at_points(ph::PointEvalHandler, proj::L2Projector, dof_values::Vector{T}) where T

Return a Vector{T} (for a 1-dimensional field) or a Vector{Vec{fielddim, T}} (for a vector field) with the field values of field fieldname in the points of the PointEvalHandler. The fieldname can be omitted if only one field is stored in dh. The field values are computed based on the dof_values and interpolated to the local coordinates by the function interpolation of the corresponding field stored in the AbstractDofHandler or the L2Projector.

Points that could not be found in the domain when constructing the PointEvalHandler will have NaNs for the corresponding entries in the output vector.

Ferrite.PointValues — Type

PointValues(cv::CellValues)
PointValues(ip_f::Interpolation, ip_g::Interpolation=ip_f)

Similar to CellValues but with a single updateable "quadrature point". PointValues are used for evaluation of functions/gradients in arbitrary points of the domain together with a PointEvalHandler.

PointValues can be created from CellValues, or from the interpolations directly.

PointValues are reinitialized like other CellValues, but since the local reference coordinate of the "quadrature point" changes this needs to be passed to reinit!, in addition to the element coordinates: reinit!(pv, coords, local_coord). Alternatively, it can be reinitialized with a PointLocation when iterating a PointEvalHandler with a PointIterator.

For function/gradient evaluation, PointValues are used in the same way as CellValues, i.e. by using function_value, function_gradient, etc, with the exception that there is no need to specify the quadrature point index (since PointValues only have 1, this is the default).

Ferrite.PointIterator — Type

PointIterator(ph::PointEvalHandler)

Create an iterator over the points in the PointEvalHandler. The elements of the iterator are either a PointLocation, if the corresponding point could be found in the grid, or nothing, if the point was not found.

A PointLocation can be used to query the cell ID with the cellid function, and can be used to reinitialize PointValues with reinit!.

Examples

ph = PointEvalHandler(grid, points)

for point in PointIterator(ph)
    point === nothing && continue # Skip any points that weren't found
    reinit!(pointvalues, point)   # Update pointvalues
    # ...
end

Ferrite.PointLocation — Type

PointLocation

Element of a PointIterator, typically used to reinitialize PointValues. Fields:

cid::Int: ID of the cell containing the point

local_coord::Vec: the local (reference) coordinate of the point

coords::Vector{Vec}: the coordinates of the cell

Ferrite.evaluate_at_grid_nodes — Function

evaluate_at_grid_nodes(dh::AbstractDofHandler, u::Vector{T}, fieldname::Symbol) where T

Evaluate the approximated solution for field fieldname at the node coordinates of the grid given the Dof handler dh and the solution vector u.

Return a vector of length getnnodes(grid) where entry i contains the evalutation of the approximation in the coordinate of node i. If the field does not live on parts of the grid, the corresponding values for those nodes will be returned as NaNs.

VTK Export

Ferrite.VTKFile — Type

Ferrite.VTKFile(filename::AbstractString, grid::AbstractGrid; kwargs...)

Create a Ferrite.VTKFile that contains an unstructured VTK grid. The keyword arguments are forwarded to WriteVTK.vtk_grid, see Data Formatting Options

This file handler can be used to to write data with

write_solution

write_celldata

write_projection

Ferrite.write_nodedata.

Ferrite.write_cellset

Ferrite.write_nodeset

Ferrite.write_dirichlet

It is necessary to call close(::VTKFile) to save the data after writing to the file handler. Using the supported do-block does this automatically: ```julia VTKFile(filename, grid) do vtk writesolution(vtk, dh, u) writecelldata(vtk, grid, celldata) end

Ferrite.write_solution — Function

write_solution(vtk::VTKFile, dh::AbstractDofHandler, u::Vector, suffix="")

Save the values at the nodes in the degree of freedom vector u to the stream. Each field in dh will be saved separately, and suffix can be used to append to the fieldname.

u can also contain tensorial values, but each entry in u must correspond to a degree of freedom in dh, see write_nodedata for details. Use write_nodedata directly when exporting values that are already sorted by the nodes in the grid.

Ferrite.write_projection — Function

write_projection(vtk::VTKFile, proj::L2Projector, vals::Vector, name::AbstractString)

Project vals to the grid nodes with proj and save to the stream.

Ferrite.write_celldata — Function

write_celldata(vtk::VTKFile, grid::AbstractGrid, celldata::AbstractVector, name::String)

Write the celldata that is ordered by the cells in the grid to the vtk file.

Ferrite.write_nodedata — Function

write_nodedata(vtk::VTKFile, grid::AbstractGrid, nodedata::Vector{Real}, name)
write_nodedata(vtk::VTKFile, grid::AbstractGrid, nodedata::Vector{<:AbstractTensor}, name)

Write the nodedata that is ordered by the nodes in the grid to the vtk stream.

When nodedata contains Tensors.Vecs, each component is exported. Two-dimensional vectors are padded with zeros.

When nodedata contains second order tensors, the index order, [11, 22, 33, 23, 13, 12, 32, 31, 21], follows the default Voigt order in Tensors.jl.

Ferrite.write_cellset — Function

write_cellset(vtk, grid::AbstractGrid)
write_cellset(vtk, grid::AbstractGrid, cellset::String)
write_cellset(vtk, grid::AbstractGrid, cellsets::Union{AbstractVector{String},AbstractSet{String})

Write all cell sets in the grid with name according to their keys and celldata 1 if the cell is in the set, and 0 otherwise. It is also possible to only export a single cellset, or multiple cellsets.

Ferrite.write_nodeset — Function

write_nodeset(vtk::VTKFile, grid::AbstractGrid, nodeset::String)

Write nodal values of 1 for nodes in nodeset, and 0 otherwise

Ferrite.write_dirichlet — Function

write_dirichlet(vtk::VTKFile, ch::ConstraintHandler)

Saves the dirichlet boundary conditions to a vtkfile. Values will have a 1 where bcs are active and 0 otherwise

Ferrite.write_cell_colors — Function

write_cell_colors(vtk::VTKFile, grid::AbstractGrid, cell_colors, name="coloring")

Write cell colors (see create_coloring) to a VTK file for visualization.

In case of coloring a subset, the cells which are not part of the subset are represented as color 0.