Grid

Ferrite.full_neighborhood — Function

full_neighborhood(grid::Grid{dim,C,T,ExclusiveTopology}, cellidx::CellIndex, include_self=false)
full_neighborhood(grid::Grid{dim,C,T,ExclusiveTopology}, faceidx::FaceIndex, include_self=false)
full_neighborhood(grid::Grid{dim,C,T,ExclusiveTopology}, vertexidx::VertexIndex, include_self=false)
full_neighborhood(grid::Grid{dim,C,T,ExclusiveTopology}, EdgeIndex::EdgeIndex, include_self=false)

Returns all directly connected entities of the same type, i.e. calling the function with a VertexIndex will return a list of directly connected vertices (connected via face/edge). If include_self is true, the given *Index is included in the returned list.

Warning: this feature is highly experimental and very likely subjected to interface changes in the future.

Ferrite.faceskeleton — Function

faceskeleton(grid) -> Vector{FaceIndex}

Returns an iterateable face skeleton. The skeleton consists of FaceIndex that can be used to reinit FaceValues.